+++ src/Compiler/Compiler.cpp
@@ -81,7 +81,7 @@ namespace Ark
                 {
                     m_bytecode.push_back(Instruction::NUMBER_TYPE);
                     auto n = std::get<BigNum>(val.value);
-                    std::string t = n.toString();
+                    std::string t = n.toString(/* base */ 16);
                     for (std::size_t i=0; i < t.size(); ++i)
                         m_bytecode.push_back(t[i]);
                 }